+++ samplesheets/assayapps/meta_ms/plugins.py
@@ -99,28 +99,18 @@ class SampleSheetAssayPlugin(SampleSheetAssayPluginPoint):
 
         return row
 
-    def get_extra_table(self, table, assay):
+    def get_shortcuts(self, assay):
         """
-        Return data for an extra content/shortcut table.
+        Return assay iRODS shortcuts.
 
-        :param table: Full table with headers (dict returned by
-                      SampleSheetTableBuilder)
         :param assay: Assay object
-        :return: Dict or None
+        :return: List or None
         """
         assay_path = self.get_assay_path(assay)
-
-        ret = {
-            'schema': {'title': 'Shortcuts for Metabolite Profiling'},
-            'cols': [
-                {'field': 'directory', 'type': 'label', 'title': 'Directory'},
-                {'field': 'links', 'type': 'irods_buttons', 'title': 'Links'},
-            ],
-            'rows': [
-                {
-                    'directory': {'value': 'Raw Data'},
-                    'links': {'path': assay_path + '/' + self.raw_data_coll},
-                }
-            ],
-        }
-        return ret
+        return [
+            {
+                'id': 'raw_data',
+                'label': 'Raw Data',
+                'path': assay_path + '/' + self.raw_data_coll,
+            }
+        ]